[SofaKernel] Split the content of DataTracker.h in different headers files
