+++ SofaKernel/modules/SofaCore/SofaCore_test/TrackedData_test.cpp
@@ -21,6 +21,7 @@
 ******************************************************************************/
 #include <sofa/core/objectmodel/BaseObject.h>
 #include <sofa/core/DataTracker.h>
+#include <sofa/core/DataTrackerFunctor.h>
 #include <sofa/simulation/AnimateBeginEvent.h>
 #include <sofa/helper/cast.h>
